fix: change directly and routing of registration to specify that it is used for users
