+++ config/routes.rb
@@ -2,7 +2,7 @@ Rails.application.routes.draw do
   namespace :api do
     namespace :v1 do 
       mount_devise_token_auth_for 'User', at: 'users' , controllers: {
-        registrations: 'api/v1/registrations'
+        registrations: 'api/v1/users/registrations'
       }
     end
   end
